14.zamiana klasy Node na prywatną i przeniesienie do klasy IntLinkedList
