+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -37,4 +37,36 @@ public class IntLinkedList {
         return ret;
     }
 
+    private class Node {
+        private int value;
+        private Node next;
+        private Node prev;
+
+        public Node(int i) {
+            value = i;
+        }
+
+        public int getValue() {
+            return value;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+
+
+        public Node getNext() {
+            return next;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
+
+
+    }
 }
